A few compiler warnings and compatiblity with GCC 4.4 hash maps:
As of gcc 4.4 they recommend using the default hasher for std::string instead
of providing a custom one as done earlier. This check-in uses a configure
variable/check to selectively include a string hasher in PEACE depending on
the compiler version and supported platform-specific configurations.

This change needs to be verified on different platforms with different
compilers etc. It should not affect runtime functionality in any way.

+++ peace/src/ParameterSetManager.cpp
@@ -77,7 +77,7 @@ ParameterSetManager::~ParameterSetManager() {
 ParameterSet*
 ParameterSetManager:: getParameterSet(const int seq1Len, const int seq2Len) {
     int minLength = std::min(seq1Len, seq2Len);
-    int setNum = 0;
+    size_t setNum = 0;
     for (setNum = 0; setNum < parameterSets.size()-1; setNum++) {
         if (minLength <= parameterSets[setNum]->maxLength) {
             break;
